servers: remove SSL FPs, fix typo & Matrix link
I think that the SSL fingerprints are useless to list as they change so
often. I suggested listing the domains that they are valid for, but I
think there is time for that later.

I added missing bracket and changed the link from matrix.org to list of
unofficial public homeservers.